requirements.txt
aiofiles==22.1.0
aiosqlite==0.17.0
asyncpg==0.27.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
dictdiffer==0.9.0
httptools==0.5.0
iso8601==1.1.0
multidict==6.0.2
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
PyJWT==2.6.0
pypika-tortoise==0.1.6
pytz==2022.5
sanic==22.9.0
sanic-routing==22.8.0
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
ujson==5.5.0
uvloop==0.17.0
websockets==10.4
